refactor: 리포트 조회 쿼리 & 비즈니스 로직 변경

See merge request s11-final/S11P31A408!85

+++ backend/api/src/main/java/com/scare/api/report/service/walk/dto/WalkingOverviewProjection.java
@@ -7,12 +7,12 @@ import lombok.Getter;
 @Getter
 public class WalkingOverviewProjection {
 
-	private Long totalWalkingTime;
+	private Integer totalWalkingTime;
 	private Long totalWalkingCnt;
 	private Double avgStressChange;
 
 	@QueryProjection
-	public WalkingOverviewProjection(Long totalWalkingTime, Long totalWalkingCnt, Double avgStressChange) {
+	public WalkingOverviewProjection(Integer totalWalkingTime, Long totalWalkingCnt, Double avgStressChange) {
 		this.totalWalkingTime = totalWalkingTime;
 		this.totalWalkingCnt = totalWalkingCnt;
 		this.avgStressChange = avgStressChange;
